Rework plugin: - use ABC for class inheritance - move Orange utils functions to decicated abstract Orange provider - better authenticated request management - move caching logic to utils
