+++ doc/README.fr.md
@@ -8,6 +8,11 @@ _Cet addon n'est pas officiellement supporté par Orange. Tous les produits, log
 ## Description
 Cet addon ajoute la TV d'Orange à Kodi. Toutes les chaînes inclues dans votre abonnement sont maintenant directement accessibles depuis Kodi !
 
+Cet addon inclue :
+- la télé en direct dans Kodi TV (en utilisant [IPTV Simple PVR](https://github.com/kodi-pvr/pvr.iptvsimple))
+- les programmes en replay
+- l'accès aux chaînes payantes qui font partie de votre souscription (pour le direct et le replay)
+
 ## Installation
 Téléchargez l'archive de la [dernière version](https://github.com/BreizhReloaded/plugin.video.orange.fr/releases/latest) et installez-la directement depuis Kodi (Settings > Addons > Install from ZIP). Installez ensuite les dépendences depuis l'écan de l'adddon (My addons > Video addons > Orange TV France > Dependencies).
 
@@ -24,6 +29,5 @@ N'oubliez pas de configurer IPTV Manager correctement et de sélectionner le bon
 | 🇫🇷        | Orange Caraïbe    |  |
 | 🇫🇷        | Orange Réunion    | merci à [@starmate](https://github.com/starmate) et [@sae-gfc](https://github.com/sae-gfc)
 
-## Limites connues
-- Cet addon ne propose pas le replay TV. Pour cela, vous pouvez aller voir du côté du très bon [Catch-up TV & More](https://github.com/Catch-up-TV-and-More/plugin.video.catchuptvandmore/)
+## Limitations connues
 - Qualité vidéo limitée à 720p en raison de limitations techniques du côté d'Orange (ils ne proposent pas plus sur leur plateforme web)
